Working compile with stub Vulkan
