+++ windowing/win32.gl.go
@@ -3,9 +3,10 @@
 package windowing
 
 /*
-#include "windowing.h"
+#include "win32.h"
 */
 import "C"
+import "unsafe"
 
 func createWindowContext(handle unsafe.Pointer, evtSharedMem *evtMem) {
 	C.window_create_gl_context(handle, evtSharedMem.AsPointer(), evtSharedMemSize)
